requirements.txt
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
joblib==1.3.2
dvc[s3]==3.20.0
boto3==1.28.25
pyyaml==6.0.1
